requirements.txt
PyYAML==6.0.2
tqdm==4.67.1
websockets==14.2
